+++ native-schema-provider/src/main/java/io/confluent/kafka/schemaregistry/builtin/NameValidator.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.confluent.kafka.schemaregistry.cflt.types;
+package io.confluent.kafka.schemaregistry.builtin;
 
 public interface NameValidator {
